+++ piezo_web_app/PiezoWebApp/src/services/validator/argument_validation_service.py
@@ -1,4 +1,4 @@
-from PiezoWebApp.src.services.validator.spark_job_property import SparkJobProperty
+from PiezoWebApp.src.services.spark_job.spark_job_property import SparkJobProperty
 from PiezoWebApp.src.services.validator.validation_result import ValidationResult
